move client and config to models
