+++ cibo/actions/commands/close.py
@@ -3,7 +3,6 @@
 from typing import List, Tuple
 
 from cibo.actions.__action__ import Action
-from cibo.client import Client
 from cibo.exception import (
     ActionMissingArguments,
     ClientNotLoggedIn,
@@ -14,6 +13,7 @@ from cibo.exception import (
     ExitNotFound,
     RoomNotFound,
 )
+from cibo.models.client import Client
 from cibo.models.message import Message, MessageRoute
