+++ cibo/actions/commands/get.py
@@ -5,7 +5,6 @@ inventory.
 from typing import List, Tuple
 
 from cibo.actions.__action__ import Action
-from cibo.client import Client
 from cibo.exception import (
     ActionMissingArguments,
     ClientNotLoggedIn,
@@ -13,6 +12,7 @@ from cibo.exception import (
     ItemNotFound,
     RoomItemNotFound,
 )
+from cibo.models.client import Client
 from cibo.models.data.item import Item
 from cibo.models.message import Message, MessageRoute
