+++ cibo/actions/commands/login.py
@@ -4,13 +4,13 @@ from typing import List, Tuple
 
 from cibo.actions.__action__ import Action
 from cibo.actions.commands.look import Look
-from cibo.client import Client
 from cibo.exception import (
     ClientIsLoggedIn,
     PasswordIncorrect,
     PlayerNotFound,
     PlayerSessionActive,
 )
+from cibo.models.client import Client
 from cibo.models.data.player import Player
 from cibo.models.message import Message, MessageRoute
