+++ cibo/actions/commands/logout.py
@@ -5,8 +5,8 @@ from typing import List, Tuple
 
 from cibo.actions.__action__ import Action
 from cibo.actions.connect import Connect
-from cibo.client import Client
 from cibo.exception import ClientNotLoggedIn
+from cibo.models.client import Client
 from cibo.models.message import Message, MessageRoute
