+++ cibo/actions/commands/quit.py
@@ -4,8 +4,8 @@ from time import sleep
 from typing import List, Optional, Tuple
 
 from cibo.actions.__action__ import Action
-from cibo.client import Client
 from cibo.exception import ClientIsLoggedIn
+from cibo.models.client import Client
 from cibo.models.message import Message, MessageRoute
