+++ cibo/actions/commands/register.py
@@ -5,8 +5,8 @@ from typing import List
 from marshmallow import ValidationError
 
 from cibo.actions.__action__ import Action
-from cibo.client import Client
 from cibo.exception import ClientIsLoggedIn, PlayerAlreadyExists, PlayerNotFound
+from cibo.models.client import Client
 from cibo.models.data.player import Player, PlayerSchema
 from cibo.models.message import Message
