+++ cibo/actions/connect.py
@@ -5,7 +5,7 @@ from typing import List, Optional
 from rich.panel import Panel
 
 from cibo.actions.__action__ import Action
-from cibo.client import Client
+from cibo.models.client import Client
 from cibo.models.message import Message
