+++ cibo/actions/error.py
@@ -3,7 +3,7 @@
 from typing import List, Optional
 
 from cibo.actions.__action__ import Action
-from cibo.client import Client
+from cibo.models.client import Client
 from cibo.models.message import Message
