+++ cibo/models/server_config.py
@@ -1,7 +1,3 @@
-"""The configuration object which supplies all the necessary objects that the
-server needs to function, and to process actions and events.
-"""
-
 from dataclasses import dataclass
 
 from cibo.resources.world import World
