+++ tests/actions/commands/test_finalize.py
@@ -1,4 +1,4 @@
-from cibo.client import ClientLoginState
+from cibo.models.client import ClientLoginState
 from cibo.models.data.player import Player
 from tests.conftest import FinalizeActionFactory
