+++ tests/actions/commands/test_login.py
@@ -1,4 +1,4 @@
-from cibo.client import ClientLoginState
+from cibo.models.client import ClientLoginState
 from cibo.output import Announcement
 from tests.conftest import LoginActionFactory
