+++ tests/actions/scheduled/test_every_minute.py
@@ -1,4 +1,4 @@
-from cibo.client import ClientLoginState
+from cibo.models.client import ClientLoginState
 from tests.conftest import EveryMinuteActionFactory
